Names components refactor
