+++ love/src/components/MainTel/MTCameraSummaryDetail/MTCameraSummaryDetail.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import styles from './MTCameraSummary.module.css';
+import styles from './MTCameraSummaryDetail.module.css';
 import StatusText from '../../GeneralPurpose/StatusText/StatusText';
 import GearIcon from '../../icons/GearIcon/GearIcon';
 import { getCameraStatusStyle } from '../../../Config';
